Memory for the dynamics import and export state now comes from MPAS pools
Previously, the memory for fields in the dynamics import and export state was
locally allocated by CAM. Now, the members of the import and export state point
to arrays that are allocated and managed in pools by MPAS-Atmosphere.

The changes to CAM-MPAS are supported by a new MPAS tag, v6.1-cam.013, that
defines new pools (dyn_in and dyn_out) in the MPAS-Atmosphere Registry file.