Redefine CAM-MPAS dynamics import/export states to match dycore prognostic vars
The CAM-MPAS dynamics import and export states were previously defined in terms
of variables that were not directly prognosed by the MPAS-Atmosphere dycore.

This commit updates the MPAS-Atmosphere externals to tag v6.1-cam.014, which
removes the old pools for import/export state in the MPAS Registry and
which defines new fields to hold the total tendencies from physics for dycore-
prognosed variables.

To match the new definitions of the 'dyn_import_t' and 'dyn_export_t' types,
the d_p_coupling routine has been modified to use the new members of
dyn_export_t, but only in the case that local_dp_map is .false..
The p_d_coupling routine has not yet been updated to the new 'dyn_import_t'
members.

The code that had been added in d_p_coupling to set a fake dycore state has
been moved to the read_inidat routine.

Lastly, there are still aspects of the use of the new dycore import and export
states that need to be corrected. At present, the vertical coordinate metric
d(zeta)/dz (in the array zz) is not valid, so the 'rho_zz' field cannot be
computed correctly in cases where there is terrain in the model simulation,
and the interpolation weight fields 'fzm' and 'fzp', used to interpolate from
layer midpoints to layer interfaces, do not contain valid values.