Read all time-invariant fields for MPAS-A in dyn_grid_init
The setup of all time-invariant fields -- including the vertical coordinate
surfaces, vertical coordinate metrics, and vertical interpolation weights --
now happens in the dyn_grid_init routine.

Because the vertical coordinate surfaces in MPAS-A are a function of
the terrain, and the vertical coordinate surfaces are read from the initial
file, there is no longer a need to read a separate terrain field, and so
the read_phis routine has been deleted.

Also, the reading of horizontal grid coordinates has been combined with
the reading of other time-invariant fields, and so the cam_mpas_read_geometry
routine has also been deleted.