Add new routine cam_mpas_cell_to_edge_winds to cam_mpas_subdriver
This commit adds a new routine to the cam_mpas_subdriver module to be used in
setting up initial conditions for horizontal velocity and in transforming
horizontal velocity tendencies from physics to dynamics. The routine projects
cell-centered wind vectors onto the unit normal vectors at edges. In support of
the use of this routine, several new variables have been added to the dynamics
import and export states.