Implement computation of dry hydrostatic pressure for dycore export state
This commit modifies which pressure fields (if any) are stored in the dycore
import and export states, and it adds a new routine for computing dry
hydrostatic pressure from the MPAS export state. Now, there are no longer any
pressure fields in dynamics import state, and dynamics export state contains
pmiddry and pintdry, which are computed at the beginning of d_p_couping by
a new routine dry_hydrostatic_pressure.