Fix segfault at line 376 of mpas_atm_core.F
Depending on compiler (and perhaps the compile or run environment), the code
in tag mpas045_cam6_2_001 would fail with a segfault at line 376 of
mpas_atm_core.F:

374 !$OMP PARALLEL DO
375       do thread=1,nThreads
376          if (.not. config_do_restart .or. (config_do_restart .and. config_do_DAcycling)) then
377             call atm_init_coupled_diagnostics(state, 1, diag, mesh, block % configs, &
378                                               cellThreadStart(thread), cellThreadEnd(thread), &

The source of the failure was the use of an unassociated pointer to
config_do_DAcycling, which was a consequence of the fact that this option
was not expected to be used by CAM-MPAS and was therefore not defined in
the "configs" pool.

The fix to this issue is to define the config_do_DAcycling option in the configs
pool with a value of .false.