Check that dycore time step evenly divides CAM time step
The dyn_run routine previously hard-wired the CAM time step as 1800 seconds,
and did not check that the specified dycore time step evenly divided this
value. Now, the CAM timestep is retrieved in every call to dyn_run via a call
to get_step_size() from the time_manager module, the dycore time step is
retrieved from an MPAS pool, the ratio of these two is verified to be
an integer, and the dycore integration time is set to the CAM time step.