Update cam_mpas_debug_stream to handle different time levels
The cam_mpas_debug_stream routine previously saved the first time level for
fields with multiple time levels to its output file. With the changes in this
commit, an optional argument, timeLevel, may be used to specify which time
level to write. This is helpful when debugging parts of the model that modify
different time levels of prognostic fields.