Use consistent values for constants in MPAS-A dycore
The MPAS dycore can now use the same values for physical constants as the rest
of CAM by obtaining these values directly from CAM's physconst module. The use
of spatially varying constants (e.g., cpairv) is still not supported, as this
will require more extensive modifications to the MPAS-A dycore.