Progress towards restart capability for MPAS-A dycore
Implement changes to enable the MPAS-Atmosphere dycore to write its restart
state to CAM's restart file, and to read its restart state from CAM's restart
file. As of this commit, restarts are not bit-for-bit: there are differences
that seem to be first arising in the Q field on the first timestep after
a restart, but further investigation is needed.