Move dry air mass scaling to new function and clean-up
This commit moves the dry air mass scaling changes inplmented in the proceeding
commits to a new subroutine, set_dry_mass. This is to be congruent with other
dycores (SE, FV3), which also have a similarly named function.

This commit also removes unsued code and adds in additional logging messages
similar to other dycores.