Update MPAS dycore set_dry_mass to physconst reference dry mass pressure
This commit updates the set_dry_mass subroutine in the MPAS dycore to use the
CESM reference pressure added in the previous commit. Depending on if there is
topography, a different reference pressure is used and passed into
set_dry_mass.

The reference pressure is then used calcuated a scaling ratio which is used to
rescale the dry air mass.