Add more log messages to set_dry_mass and compute scaled surface pressure
This commit adds additional log messages to the set_dry_mass routine and
rewrites previous log messages to be similar to other dycores.

In one of the log messages above, it logs the average dry surface
pressure after it is scaled, the computation of which is added in this commit.
This is similar to the output of other dycores.