Fix restart failures due to unallocated physics tendency fields
The tend_ru_physics, tend_rtheta_physics, and tend_rho_physics fields are
allocated in the call to mpas_atm_dynamics_init. For initial runs, these
tendency fields are not read from the initial file; however, these tendencies
were read from restart files before the call to mpas_atm_dynamics_init, leading
to model failure. Since the tend_ru_physics, tend_rtheta_physics, and
tend_rho_physics field are not required in the MPAS-A dycore restart state for
exact restarts (because they are recomputed in the first call to p_d_coupling
before the first dynamics step), these fields can simply be removed from the set
of fields to be written to and read from restart files.

Also included in this commit is a one-line change to initialize domainID after
the allocation of the domain_ptr.