Add mpasa60 topography and initial conditions to namelist defaults
This commits adds the mpasa60 topography (bnd_topo) file to the namelist
defaults file. As well, it adds references to the initial conditions.

NOTE: Currently, the mpasa60 grids might not work. It appears that there is
currently an error/issue with the mpasa60 topography files, which cause the
model to encouter an NaN in the W field after the first timestep.