Remove online land type option when using GEOS-Chem and use Olson/XLAI
Dry deposition in GEOS-Chem now uses a combination of CLM dry deposition
velocities and GEOS-Chem computed velocities which is determined by
presence of ocean (100% CLM velocities if no ocean, 100% GEOS-Chem
velocities if all ocean, and a scaled value of the two if in between).
This update also uncomments several code that set State_Met fields, and
the logicals for CLM drydep vel and online land type is removed (the
former is effectively now always true and the latter is always false).

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>