Update maximum CAM history field length to 32 (up from 24).
This updates cam_history_support's fieldname_len parameter from 24 to 32, in order to accommodate longer names for history outputs. This is in support of implementing GEOS-Chem into CESM, as we recently implemented a feature to automatically pass GEOS-Chem diagnostic quantities to CAM history. Some of the GEOS-Chem field names are longer than 24 characters and thus we need to increase the maximum length.

Signed-off-by: Haipeng Lin <hplin@seas.harvard.edu>