Move short-lived reference MMR array from chem_mods to short-lived module
Reference mixing ratios for short-lived species can now be passed to
subroutine register_short_lived_species as an optional argument. If passed
then the new module variable for reference values will be allocated and
set using the values. Later during initialization these values will be
used if the module array is allocated. Also added is a new finalize
subroutine for the short_lived_species module so that the module
array is deallocated.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>