Remove NTHRDS overrides for GEOS-Chem
NTHRDS is 1 by default when using GEOS-Chem compsets so these overrides
are not necessary. We find that manually changing NTHRDS to something
other than 1 breaks the model, but this is the case with non-GEOS-Chem
compsets as well. Unless this becomes a problem the overrides are not
needed at this time.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>