Exit with warning in build-namelist if hemco_config_file or hemco_diagn_file are specified in user namelist.
Current behavior in build-namelist enforces use of config files in case directory which are copied to run directory at submit time.
User customization to user_nl_cam for the path of HEMCO config files will not take effect when running with GEOS-Chem chemistry, instead the user should edit or symlink from the files in the case directory similar to standalone GEOS-Chem operation.
This commit adds a check in build-namelist for this customization and aborts with error to warn user that this is unsupported.

Signed-off-by: Haipeng Lin <hplin@seas.harvard.edu>