Revert dep_data_file input_pathname to "abs"; use new file for geoschem
Previously the dep_data_file default for GEOS-Chem was not actually being
used due to inclusion in the use case files. This let a bug go under the
radar that dep_data_file input_pathname="abs" was removed from the
namelist definitions, meaning the configured inputdata path would not be
used as a prefix to whatever dep_data_file was set as. This allowed
GEOS-Chem compsets to work since the full path was in the use case files,
but it resulted in non-GEOS-Chem compsets failing due to an incomplete
path for the deposition netcdf file containing Henry's Law coeffs et al.

This update restores using inputdata path as a prefix to dep_data_file,
and so fixes the non-GEOS-Chem compsets. It also removes the setting of
dep_data_file in the GEOS-Chem compsets and updates the filename in
the namelist defaults for dep_data_file if using GEOS-Chem.

GEOS-Chem compsets will not work until someone at NCAR moves the file to
the target inputdata directory. The request to do this has been made.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>