Wrap MPAS longitude values to [0,2pi) range
This commit addresses issues that were seen when using ncdata with
longitudes in the [-pi, pi] range in cases with full physics.  This
would cause the atmosphere to become unphysical and runs to fail. This
change is made just after the longitude values are read to ensure they
are correct for the remainder of execution.